od5.0: Fixed some conversion warnings
Summary:
src\Network\tests\netreq.cc(32) : warning C4244: 'argument' : conversion from 'int' to 'unsigned short', possible loss of data
plugins\uiHorizonAttrib\uicontourtreeitem.cc(434) : warning C4244: '=' : conversion from 'double' to 'float', possible loss of data

Test Plan: compiles

Reviewers: nageswara

Reviewed By: nageswara

Differential Revision: https://phabricator.dgbes.com:8443/D6488

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@37456 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/Network/tests/netreq.cc
@@ -29,7 +29,7 @@ public:
 
     bool runTest( bool sendkill )
     {
-	Network::RequestConnection conn( hostname_, port_ );
+	Network::RequestConnection conn( hostname_, (unsigned short)port_ );
 	mAttachCB( conn.packetArrived, Tester::packetArrivedCB );
 
 	Network::RequestPacket packet;
